+++ entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+echo "Pokretanje migracija..."
+npx prisma migrate deploy
+echo "Pokretanje aplikacije..."
+exec npm start
