+++ prisma.config.ts
@@ -1,6 +1,5 @@
 // This file was generated by Prisma, and assumes you have installed the following:
 // npm install --save-dev prisma dotenv
-import "dotenv/config";
 import { defineConfig } from "prisma/config";
 
 export default defineConfig({
